requirements.txt
lib-platform
requests
subprocess.run
regex
psutil
times
datetime
sockets
configparser
requires.io
argparse
typing
docker
pathlib
pyyaml
paramiko
flet
pysmb
schedule
croniter